chore(FDS-250) Loading/Empty states, tests
